requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
wavelink==3.3.0
pytz==2024.1 